Set state.bot property to "Copy always"
This is critical for having the bot work when it is deployed to Azure. Without it, "Test in Web Chat" hangs forever with "Waiting for bot to be ready". The other projects in this group of dotnet_core bot samples are correct. This is the only one that lacks the setting.
Best, Bruce.